refactor: move plugin to scripts dir
